ai_engine/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
jinja2
sqlalchemy
python-multipart
PyJWT
passlib
bcrypt
websockets